+++ sdk/src/userMap/userStatsMap.ts
@@ -4,7 +4,6 @@ import {
 	OrderRecord,
 	UserStatsAccount,
 	UserStats,
-	UserStatsSubscriptionConfig,
 	WrappedEvent,
 	DepositRecord,
 	FundingPaymentRecord,
@@ -14,12 +13,12 @@ import {
 	NewUserRecord,
 	LPRecord,
 	InsuranceFundStakeRecord,
-	StateAccount,
+	BulkAccountLoader,
+	PollingUserStatsAccountSubscriber,
 } from '..';
-import { AccountInfo, PublicKey } from '@solana/web3.js';
+import { PublicKey } from '@solana/web3.js';
 
 import { UserMap } from './userMap';
-import { Buffer } from 'buffer';
 
 export class UserStatsMap {
 	/**
@@ -27,39 +26,45 @@ export class UserStatsMap {
 	 */
 	private userStatsMap = new Map<string, UserStats>();
 	private driftClient: DriftClient;
-	private accountSubscription: UserStatsSubscriptionConfig;
-	private lastNumberOfAuthorities;
-	private syncCallback = async (state: StateAccount) => {
-		if (state.numberOfAuthorities !== this.lastNumberOfAuthorities) {
-			await this.sync();
-			this.lastNumberOfAuthorities = state.numberOfAuthorities;
-		}
-	};
+	private bulkAccountLoader: BulkAccountLoader;
 
-	constructor(
-		driftClient: DriftClient,
-		accountSubscription: UserStatsSubscriptionConfig
-	) {
+	/**
+	 * Creates a new UserStatsMap instance.
+	 *
+	 * @param {DriftClient} driftClient - The DriftClient instance.
+	 * @param {BulkAccountLoader} [bulkAccountLoader] - If not provided, a new BulkAccountLoader with polling disabled will be created.
+	 */
+	constructor(driftClient: DriftClient, bulkAccountLoader?: BulkAccountLoader) {
 		this.driftClient = driftClient;
-		this.accountSubscription = accountSubscription;
+		if (!bulkAccountLoader) {
+			bulkAccountLoader = new BulkAccountLoader(
+				driftClient.connection,
+				driftClient.opts.commitment,
+				0
+			);
+		}
+		this.bulkAccountLoader = bulkAccountLoader;
 	}
 
-	public async subscribe() {
+	public async subscribe(authorities: PublicKey[]) {
 		if (this.size() > 0) {
 			return;
 		}
 
 		await this.driftClient.subscribe();
-		this.lastNumberOfAuthorities =
-			this.driftClient.getStateAccount().numberOfAuthorities;
-		this.driftClient.eventEmitter.on('stateAccountUpdate', this.syncCallback);
-
-		await this.sync();
+		await this.sync(authorities);
 	}
 
+	/**
+	 *
+	 * @param authority that owns the UserStatsAccount
+	 * @param userStatsAccount optional UserStatsAccount to subscribe to, if undefined will be fetched later
+	 * @param skipFetch if true, will not immediately fetch the UserStatsAccount
+	 */
 	public async addUserStat(
 		authority: PublicKey,
-		userStatsAccount?: UserStatsAccount
+		userStatsAccount?: UserStatsAccount,
+		skipFetch?: boolean
 	) {
 		const userStat = new UserStats({
 			driftClient: this.driftClient,
@@ -67,9 +72,18 @@ export class UserStatsMap {
 				this.driftClient.program.programId,
 				authority
 			),
-			accountSubscription: this.accountSubscription,
+			accountSubscription: {
+				type: 'polling',
+				accountLoader: this.bulkAccountLoader,
+			},
 		});
-		await userStat.subscribe(userStatsAccount);
+		if (skipFetch) {
+			await (
+				userStat.accountSubscriber as PollingUserStatsAccountSubscriber
+			).addToAccountLoader();
+		} else {
+			await userStat.subscribe(userStatsAccount);
+		}
 
 		this.userStatsMap.set(authority.toString(), userStat);
 	}
@@ -77,7 +91,7 @@ export class UserStatsMap {
 	public async updateWithOrderRecord(record: OrderRecord, userMap: UserMap) {
 		const user = await userMap.mustGet(record.user.toString());
 		if (!this.has(user.getUserAccount().authority.toString())) {
-			await this.addUserStat(user.getUserAccount().authority);
+			await this.addUserStat(user.getUserAccount().authority, undefined, false);
 		}
 	}
 
@@ -156,9 +170,19 @@ export class UserStatsMap {
 		return this.userStatsMap.get(authorityPublicKey);
 	}
 
+	/**
+	 * Enforce that a UserStats will exist for the given authorityPublicKey,
+	 * reading one from the blockchain if necessary.
+	 * @param authorityPublicKey
+	 * @returns
+	 */
 	public async mustGet(authorityPublicKey: string): Promise<UserStats> {
 		if (!this.has(authorityPublicKey)) {
-			await this.addUserStat(new PublicKey(authorityPublicKey));
+			await this.addUserStat(
+				new PublicKey(authorityPublicKey),
+				undefined,
+				false
+			);
 		}
 		return this.get(authorityPublicKey);
 	}
@@ -171,53 +195,24 @@ export class UserStatsMap {
 		return this.userStatsMap.size;
 	}
 
-	public async sync() {
-		const programAccounts =
-			await this.driftClient.connection.getProgramAccounts(
-				this.driftClient.program.programId,
-				{
-					commitment: this.driftClient.connection.commitment,
-					filters: [
-						{
-							memcmp:
-								this.driftClient.program.coder.accounts.memcmp('UserStats'),
-						},
-					],
-				}
-			);
-
-		const programAccountMap = new Map<string, AccountInfo<Buffer>>();
-		for (const programAccount of programAccounts) {
-			programAccountMap.set(
-				new PublicKey(programAccount.account.data.slice(8, 40)).toString(),
-				programAccount.account
-			);
-		}
-
-		for (const key of programAccountMap.keys()) {
-			if (!this.has(key)) {
-				const userStatsAccount =
-					this.driftClient.program.account.userStats.coder.accounts.decode(
-						'UserStats',
-						programAccountMap.get(key).data
-					);
-				await this.addUserStat(new PublicKey(key), userStatsAccount);
-			}
-		}
+	/**
+	 * Sync the UserStatsMap
+	 * @param authorities list of authorities to derive UserStatsAccount public keys from.
+	 * You may want to get this list from UserMap in order to filter out idle users
+	 */
+	public async sync(authorities: PublicKey[]) {
+		await Promise.all(
+			authorities.map((authority) =>
+				this.addUserStat(authority, undefined, true)
+			)
+		);
+		await this.bulkAccountLoader.load();
 	}
 
 	public async unsubscribe() {
 		for (const [key, userStats] of this.userStatsMap.entries()) {
 			await userStats.unsubscribe();
 			this.userStatsMap.delete(key);
-		}
-
-		if (this.lastNumberOfAuthorities) {
-			this.driftClient.eventEmitter.removeListener(
-				'stateAccountUpdate',
-				this.syncCallback
-			);
-			this.lastNumberOfAuthorities = undefined;
 		}
 	}
 }