+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 
 ### Breaking
+- sdk: UserStatsMap use bulkAccountLoader (`UserStatsMap.subscribe` and `UserStatsMap.sync` now requires list of authorities) ([#716](https://github.com/drift-labs/protocol-v2/pull/716))
 
 ## [2.47.0] - 2023-11-26
 
@@ -20,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - program: accelerated idle update for users with <$1000 equity
 
 ### Fixes
+
+- ts-sdk: fix to modify order booleans
 
 ### Breaking
 
@@ -77,14 +80,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- program: add accelerated user update idle ([#669](https://github.com/drift-labs/protocol-v2/pull/669)) 
+- program: add accelerated user update idle ([#669](https://github.com/drift-labs/protocol-v2/pull/669))
 - program: make user status a bit flag ([#619](https://github.com/drift-labs/protocol-v2/pull/619))
 - program: place and take uses auction end price for market orders ([#650](https://github.com/drift-labs/protocol-v2/pull/650))
 - program: reduce cus for place_spot_order ([#662](https://github.com/drift-labs/protocol-v2/pull/662))
 - program: bump max sub accounts to 15k
 - program: user custom margin ratio works with spot ([#633](https://github.com/drift-labs/protocol-v2/pull/633))
 - program: add swap price bands ([#611](https://github.com/drift-labs/protocol-v2/pull/611))
-- program: add 5min twap price bands to liquidate_perp and liquidate_spot ([#570](https://github.com/drift-labs/protocol-v2/pull/570)) 
+- program: add 5min twap price bands to liquidate_perp and liquidate_spot ([#570](https://github.com/drift-labs/protocol-v2/pull/570))
 - program: add positive perp funding rate offset ([#576](https://github.com/drift-labs/protocol-v2/pull/576/files))
 
 ### Fixes